clean and move api
